[DP-283] refactor: 채팅 관련 유틸리티 함수 분리 후 커스텀 훅으로 사용 #274
